Revert "fix: 修复 Proptype 的错误, 优化开发体验 (#73)"
This reverts commit e5ee243ae8cc7cd9e24dfb04bc4a22fec68ec9e0.
