+++ .travis.yml
@@ -7,7 +7,7 @@ notifications:
   - hualei5280@gmail.com
 
 node_js:
-- 10
+- 6
 
 before_install:
 - |
