+++ src/PropsType.ts
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 
 export interface Indicator {
   activate?: React.ReactNode;
@@ -7,15 +7,15 @@ export interface Indicator {
   finish?: React.ReactNode;
 }
 
-export interface IPropsType {
-  onRefresh: () => void;
-  getScrollContainer?: () => React.ReactNode;
-  direction?: 'down' | 'up';
+export interface PropsType {
+  getScrollContainer: () => React.ReactNode;
+  direction: 'down' | 'up';
   refreshing?: boolean;
-  distanceToRefresh?: number;
-  indicator?: Indicator;
+  distanceToRefresh: number;
+  onRefresh: () => void;
+  indicator: Indicator;
   prefixCls?: string;
   className?: string;
-  damping?: number;
   style?: React.CSSProperties;
+  damping: number;
 }
