--- src/StaticRenderer.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-export default class StaticRenderer extends React.Component<any, any> {
-  shouldComponentUpdate(nextProps: any) {
-    return nextProps.shouldUpdate;
-  }
-  render() {
-    return <div>{this.props.render()}</div>;
-  }
-}
