--- tslint.json
@@ -1,7 +1,0 @@
-{
-  "extends": "rc-tools/config/tslint.json",
-  "rules": {
-    "member-access": [true, "no-public"],
-    "variable-name": [false]
-  }
-}
